+++ src/sentry/toolbar/utils/csp.py
